api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-dotenv==1.4.2
django-ninja==1.3.0
django-taggit==6.1.0
django-tinymce==4.1.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
